+++ src/multi_test/mod.rs
@@ -7,5 +7,7 @@ pub mod macros;
 pub mod modules;
 mod runner;
 
+pub mod api;
+
 pub use crate::create_contract_wrappers;
 pub use runner::MultiTestRunner;